feat: update example to Next.js 13
