+++ typescript/rest-nextjs-api-routes/next.config.js
@@ -1,8 +1,4 @@
-const { withSuperjson } = require('next-superjson')
-
 /** @type {import('next').NextConfig} */
-const nextConfig = withSuperjson()({
-  reactStrictMode: true,
-})
+const nextConfig = {}
 
 module.exports = nextConfig
