+++ typescript/rest-nextjs-api-routes/package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@prisma/client": "4.14.0",
     "next": "13.4.1",
-    "next-superjson": "0.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-markdown": "8.0.7",
