+++ typescript/rest-nextjs-api-routes/src/app/create/page.tsx
@@ -1,0 +1,63 @@
+'use client'
+
+import Router from 'next/router'
+import { useState } from 'react'
+import styles from './page.module.css'
+import Link from 'next/link'
+
+export default function Draft() {
+  const [title, setTitle] = useState('')
+  const [content, setContent] = useState('')
+  const [authorEmail, setAuthorEmail] = useState('')
+
+  const submitData = async (e: React.SyntheticEvent) => {
+    e.preventDefault()
+    try {
+      const body = { title, content, authorEmail }
+      await fetch(`/api/post`, {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(body),
+      })
+      await Router.push('/drafts')
+    } catch (error) {
+      console.error(error)
+    }
+  }
+
+  return (
+    <div>
+      <form onSubmit={submitData}>
+        <h1>Create Draft</h1>
+        <input
+          autoFocus
+          onChange={(e) => setTitle(e.target.value)}
+          placeholder="Title"
+          type="text"
+          value={title}
+        />
+        <input
+          onChange={(e) => setAuthorEmail(e.target.value)}
+          placeholder="Author (email address)"
+          type="text"
+          value={authorEmail}
+        />
+        <textarea
+          cols={50}
+          onChange={(e) => setContent(e.target.value)}
+          placeholder="Content"
+          rows={8}
+          value={content}
+        />
+        <input
+          disabled={!content || !title || !authorEmail}
+          type="submit"
+          value="Create"
+        />
+        <Link className={styles.black} href="/">
+          or Cancel
+        </Link>
+      </form>
+    </div>
+  )
+}
