+++ typescript/rest-nextjs-api-routes/src/components/Header.tsx
@@ -1,35 +1,33 @@
+'use client'
+
 import React from 'react'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
+import { usePathname } from 'next/navigation'
 import styles from '@/components/Header.module.css'
 
-const Header: React.FC = () => {
-  const router = useRouter()
+export default function Header() {
+  const currentPathname = usePathname()
   const isActive: (pathname: string) => boolean = (pathname) =>
-    router.pathname === pathname
+    currentPathname === pathname
 
   return (
     <nav>
       <div className={styles.left}>
-        <Link href="/" legacyBehavior>
-          <a className={styles.bold} data-active={isActive('/')}>
-            Blog
-          </a>
+        <Link className={styles.bold} href="/" data-active={isActive('/')}>
+          Blog
         </Link>
-        <Link href="/drafts" legacyBehavior>
-          <a data-active={isActive('/drafts')}>Drafts</a>
+        <Link href="/drafts" data-active={isActive('/drafts')}>
+          Drafts
         </Link>
       </div>
       <div className={styles.right}>
-        <Link href="/signup" legacyBehavior>
-          <a data-active={isActive('/signup')}>Signup</a>
+        <Link href="/signup" data-active={isActive('/signup')}>
+          Signup
         </Link>
-        <Link href="/create" legacyBehavior>
-          <a data-active={isActive('/create')}>+ Create draft</a>
+        <Link href="/create" data-active={isActive('/create')}>
+          + Create draft
         </Link>
       </div>
     </nav>
   )
 }
-
-export default Header
