--- typescript/rest-nextjs-api-routes/src/pages/api/publish/[id].ts
@@ -1,15 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next'
-import prisma from '../../../lib/prisma'
-
-// PUT /api/publish/:id
-export default async function handle(
-  req: NextApiRequest,
-  res: NextApiResponse,
-) {
-  const postId = req.query.id
-  const post = await prisma.post.update({
-    where: { id: Number(postId) },
-    data: { published: true },
-  })
-  return res.json(post)
-}
