--- typescript/rest-nextjs-api-routes/src/pages/api/user.ts
@@ -1,16 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next'
-import prisma from '../../lib/prisma'
-
-// POST /api/user
-// Required fields in body: name, email
-export default async function handle(
-  req: NextApiRequest,
-  res: NextApiResponse,
-) {
-  const result = await prisma.user.create({
-    data: {
-      ...req.body,
-    },
-  })
-  return res.status(201).json(result)
-}
